# perf: improve connection pool performance