--- a/data-plane/gateway/datapath/Cargo.toml
+++ b/data-plane/gateway/datapath/Cargo.toml
@@ -25,8 +25,8 @@ tracing = "0.1.41"
 tracing-opentelemetry = "0.29.0"
 
 [dev-dependencies]
-tracing-test = { version = "0.2.5", features = ["no-env-filter"] }
 criterion = { version = "0.5", features = ["html_reports"] }
+tracing-test = { version = "0.2.5", features = ["no-env-filter"] }
 
 [build-dependencies]
 protoc-bin-vendored = "3.1.0"