--- a/data-plane/gateway/datapath/src/tables/pool.rs
+++ b/data-plane/gateway/datapath/src/tables/pool.rs
@@ -73,8 +73,7 @@ impl<T> Pool<T> {
         // If length is equal to capacity, resize the pool
         if self.len == self.capacity {
             // Resize the pool
-            self.pool
-                .resize_with(2 * self.capacity, || None);
+            self.pool.resize_with(2 * self.capacity, || None);
             self.bitmap.grow(self.capacity, false);
             self.capacity *= 2;